+++ phoenix/models/__init__.py
@@ -16,6 +16,16 @@ def mongodb(registry):
     settings = registry.settings
     return pymongo.Connection(settings['mongodb.url'])[settings['mongodb.db_name']]
 
+def auth_protocols(request):
+    # TODO: refactor auth settings handling
+    settings = request.db.settings.find_one()
+    protocols = ['esgf', 'openid', 'ldap', 'oauth2']
+    if settings is not None:
+        if settings.has_key('auth'):
+            if settings['auth'].has_key('protocol'):
+                protocols = settings['auth']['protocol']
+    return protocols
+
 def get_user(request):
     userid = authenticated_userid(request)
     return request.db.users.find_one(dict(identifier=userid))
@@ -68,7 +78,21 @@ def user_cert_valid(request, valid_hours=8):
             return True
     return False
 
+def load_settings(request):
+    defaults = dict(solr_maxrecords = -1, solr_depth = 2)
+    
+    settings = request.db.settings.find_one()
+    if not settings:
+        settings = save_settings(request, defaults)
+    for key in defaults.keys():
+        if not key in settings:
+            settings[key] = defaults[key]
+    return settings
 
+def save_settings(request, settings):
+    request.db.settings.save(settings)
+    return request.db.settings.find_one()
+