+++ phoenix/panels/__init__.py
@@ -1,8 +1,10 @@
 from pyramid_layout.panel import panel_config
 from pyramid.security import authenticated_userid, has_permission
+from phoenix.models import get_user
 
 import logging
 logger = logging.getLogger(__name__)
+
 
 @panel_config(name='navbar', renderer='phoenix:templates/panels/navbar.pt')
 def navbar(context, request):
@@ -25,22 +27,35 @@ def navbar(context, request):
     if has_permission('admin', request.context, request):
         subitems.append( nav_item('Settings', request.route_path('settings'), icon="fa fa-wrench") )
     
-    login = request.current_route_url() == request.route_url('account_login', protocol='oauth2')
+    login = 'login' in request.current_route_url()
 
-    return dict(title='Phoenix', items=items, subitems=subitems, username=authenticated_userid(request), login=login)
+    username = None
+    try:
+        user = get_user(request)
+        if user:
+            username = user.get('name')
+    except:
+        logger.exception('could not get username')
+
+    return dict(title='Phoenix', items=items, subitems=subitems, username=username, login=login)
+
 
 @panel_config(name='breadcrumbs', renderer='phoenix:templates/panels/breadcrumbs.pt')
 def breadcrumbs(context, request):
     lm = request.layout_manager
     return dict(breadcrumbs=lm.layout.breadcrumbs)
 
+
 @panel_config(name='sidebar', renderer='phoenix:templates/panels/sidebar.pt')
 def sidebar(context, request):
     return dict()
 
+
 @panel_config(name='footer', renderer='phoenix:templates/panels/footer.pt')
 def footer(context, request):
-    return dict(version="0.4.1")
+    from phoenix import get_version
+    return dict(version=get_version())
+
 
 @panel_config(name='headings')
 def headings(context, request):