+++ phoenix/settings/views/ldap_config.py
@@ -24,7 +24,7 @@ class Ldap(SettingsView):
             ldap_settings = dict()
 
         # Generate form
-        from phoenix.schema.settings import LdapSchema
+        from phoenix.settings.schema import LdapSchema
         ldap_form = Form(schema = LdapSchema(), buttons = ('submit',), formid = 'deform')
 
         if 'submit' in self.request.params:
@@ -42,6 +42,8 @@ class Ldap(SettingsView):
                 ldap_settings['base_dn']     = appstruct['base_dn']
                 ldap_settings['filter_tmpl'] = appstruct['filter_tmpl']
                 ldap_settings['scope']       = appstruct['scope']
+                # Optional:
+                ldap_settings['name']        = appstruct['name']
                 ldap_settings['email']       = appstruct['email']
                 self.db.ldap.save(ldap_settings)